migrate to /test/legacy_test
